requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.2
django-filter==24.3
django_csp==3.8
geopandas==1.0.1
gunicorn==23.0.0
joblib==1.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2
whitenoise==6.8.2
